add 187

+++ LeetCode/517_超级洗衣机.cpp
@@ -10,14 +10,14 @@ public:
         vector<bool> moveds(machines.size(), false);
         for (size_t i = 0; i < origin_machines.size(); i++) {
             left_sum += origin_machines[i];
-            if (left_sum < avg_num * (i + 1)) {
+            if (left_sum < avg_num * static_cast<int>(i + 1)) {
                 if (!moveds[i + 1] && origin_machines[i + 1] > 0) {
                     machines[i]++;
                     machines[i + 1]--;
                     moveds[i + 1] = true;
                 }
             }
-            else if (left_sum > avg_num * (i + 1)) {
+            else if (left_sum > avg_num * static_cast<int>(i + 1)) {
                 if (!moveds[i] && origin_machines[i] > 0) {
                     machines[i]--;
                     machines[i + 1]++;
@@ -41,9 +41,9 @@ public:
         */
         int result = 0;
         int left_sum = 0;
-        for (int i = 0; i < machines.size(); i++) {
+        for (size_t i = 0; i < machines.size(); i++) {
             left_sum += machines[i];
-            result = max(abs(left_sum - avg_num * (i + 1)), result); // 差额
+            result = max(abs(left_sum - avg_num * static_cast<int>(i + 1)), result); // 差额
             result = max(machines[i] - avg_num, result); // 一台洗衣机多的
         }
         return result;
